# openapi.yaml
openapi: 3.0.0
info:
  title: PawConnect API - Gestión de Animales
  description: Documentación de las operaciones CRUD para el modelo de Animales.
  version: 1.0.0

servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo local

paths:
  /animales:
    get:
      summary: Listar todos los animales
      responses:
        '200':
          description: Listado obtenido correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Animal'
    post:
      summary: Registrar un nuevo amiguito
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Animal'
      responses:
        '201':
          description: Animal creado exitosamente
        '400':
          description: Error de validación (faltan campos o datos incorrectos)

  /animales/{id}:
    get:
      summary: Obtener detalle de un animal por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Datos del animal devueltos
        '404':
          description: Animal no encontrado
    put:
      summary: Actualizar datos de un animal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Animal'
      responses:
        '200':
          description: Animal actualizado correctamente
        '400':
          description: Datos de actualización inválidos
    delete:
      summary: Eliminar un animal de la base de datos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Animal eliminado (sin contenido de respuesta)

components:
  schemas:
    Animal:
      type: object
      required:
        - nombre
        - especie
        - peso
        - fechaNacimiento
      properties:
        nombre:
          type: string
          example: "Boby"
        especie:
          type: string
          enum: [perro, gato, otro]
          example: "perro"
        peso:
          type: number
          example: 12.5
        fechaNacimiento:
          type: string
          format: date
          example: "2024-01-15"
        disponible:
          type: boolean
          default: true